requirements.txt
aiogram==3.7.0
supabase==2.4.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiohttp==3.9.5
